feat(processing): add patch generator module
